move fetching HTML to FagsakController
